pyproject.toml
[project]
name = "nanabanana-for-darnytsia"
version = "0.1.0"
description = "Веб-додаток NanaBanana для генерації зображень за допомогою Google Gemini API для Darnytsia"
requires-python = ">=3.14"
dependencies = [
    "streamlit>=1.5.0",
    "google-genai>=0.2.0",
    "Pillow>=10.0.0",
    "python-telegram-bot>=20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gemini_image_generator"]

